uniform to random
Just a little performance boost

+++ ga2/gaDisc.py
@@ -1,67 +1,64 @@
-# This program impliments the discrete genetic algorithm
-# here, Discrete means that the new generation is generated
-# all at the same time by considering the top agents at that
-# given point.
+"""
+This program impliments the discrete genetic algorithm
+here, Discrete means that the new generation is generated
+all at the same time by considering the top agents at that
+given point.
+"""
 
 import os
+import ga2
 import shutil
 import random
-import ga2.gaUtils.gaDeamon as deamon
+import ga2.gaUtils.gaDaemon as daemon
 
 class Session:
     """
     This is the session class which is used to access all
     the agents in the current session, invoke different
     functions on this session and setting session
-    parameters.
+    parameters. This class also holds the meta data of the
+    current session while using the 'UNSAFE' mode option
     """
 
     def __init__(self,sessID=None, agentCount=100, numParam=10, fixParam=True,
-                 spread=2, survival=0.05, genecopy=0.9, mutation=0.01, generateMode='A'):
+                 spread=2, survival=0.05, genecopy=0.9, mutation=0.01, generateMode='A',
+                 mode='UNSAFE', valType='FLOAT'):
+        
 
-        if(sessID is None):
-            self.sessID = deamon.getNewSessionID()
-            self.agentCount = agentCount
-            self.numParam = numParam
-            self.fixParam = fixParam
-            self.spread = spread
-            self.generateMode = generateMode
-            self.mutation = mutation
-            self.genecopy = genecopy
-            self.survival = survival
-            deamon.setSession(self)
-
-        else:
-            tmpObj = deamon.getSession(sessID)
-            self.sessID = tmpObj.sessID
-            self.agentCount = tmpObj.agentCount
-            self.numParam = tmpObj.numParam
-            self.fixParam = tmpObj.fixParam
-            self.spread = tmpObj.spread
-            self.mutation = mutation
-            self.survival = survival
-            self.genecopy = genecopy
-            self.generateMode = tmpObj.generateMode
+        self.sessID        = daemon.getNewSessionID()
+        self.spread        = spread
+        self.valType       = valType
+        self.numParam      = numParam
+        self.fixParam      = fixParam
+        self.mutation      = mutation
+        self.genecopy      = genecopy
+        self.survival      = survival
+        self.agentCount    = agentCount
+        self.generateMode  = generateMode
+        
+        self.currGen       = set()
+        self.agentBasket   = dict()
+        self.nextAgentID   = 0
+        self.lock          = 0
+        self.mode          = mode
+        
         
     def init(self):
         """
         INPUT       : None
-        OUTPUT      : Returns 1 on success, None on failure
+        OUTPUT      : None
 
-        DESCRIPTION : Initializes the session with initial
-                      values based on the session conf .
-
-        we create self.agentCount number of agents here
-        using the deamon function createAgent
+        DESCRIPTION : Initializes the session with agents
         """
         
         try:
             for i in range(self.agentCount):
-                agentObj = deamon.createAgent(self)
+                agentObj = daemon.createAgent(self)
         
         except Exception as e:
-            print("gaDiscrete.init failed with error as : {}".format(e))
-        
+            print("gaDisc.init failed with error as : {}".format(e))
+
+            
     def getAllAgents(self):
         """
         INPUT       : None
@@ -69,22 +66,17 @@ class Session:
                       agents in a list, or None on error
 
         DESCRIPTION : Self explanatory from OUTPUT
-
-        
-        toRet is the set containing the agentIDs of all
-        agents who are currently active.
-        toRet is populated by using the deamon function
-        getCurrGen
         """
         
         try:
-            toRet = deamon.getCurrGen(self.sessID)
+            toRet = daemon.getCurrGen(self)
         except Exception as e:
-            print("gaDiscrete.getAllAgents failed with error as : {}".format(e))
+            print("gaDisc.getAllAgents failed with error as : {}".format(e))
             toRet = None
             
         return(toRet)
 
+    
     def getAgent(self, agentID):
         """
         INPUT       : AgentID
@@ -92,37 +84,31 @@ class Session:
                       None on error
 
         DESCRIPTION : Self explanatory from OUTPUT
-
-        agentObj is the agent corresponding to the agentID = agentID
-        and belonging to the session with sessID = sessID
-        uses the deamon function
         """
 
         try:
-            agentObj = deamon.getAgent(self.sessID, agentID)
+            agentObj = daemon.getAgent(self, agentID)
             return(agentObj)
         
         except Exception as e:
-            print("gaDiscrete.getAgent failed with error as : {}".format(e))
+            print("gaDisc.getAgent failed with error as : {}".format(e))
+            return(None)
+        
         
     def updateAgent(self, agentObj):
         """
         INPUT       : Agent object
         OUTPUT      : Returns 1 on success, None on failure
 
-        DESCRIPTION : Updates the status of Agent in the cluster.
-
-        Stores the agent in the current session agent's file
-        this storage is for persistance and to save the RAM
-        for tree building or other more important jobs
-        uses the deamon storeAgent function
+        DESCRIPTION : Updates the status of Agent in the current session
         """
 
         try:
-            return(deamon.storeAgent(agentObj))
+            return(daemon.storeAgent( self, agentObj))
         
         except Exception as e:
-            print("gaDiscrete.updateAgent failed with error as : {}".format(e))            
+            print("gaDisc.updateAgent failed with error as : {}".format(e))
+            return(None)
     
     def getBestAgent(self):
         """
@@ -135,23 +121,25 @@ class Session:
         This is used to retrieve the best agent in the current
         pool of agents based on their fitness value with
         higher fitness corresponding to being better
-        uses deamon functions getCurrGen, getAgent
+        uses daemon functions getCurrGen, getAgent
         """
 
         try:
+            
             bestAgent = None
-            currGen = deamon.getCurrGen(self.sessID)
+            currGen = daemon.getCurrGen(self)
             for agentID in currGen:
-                agent = deamon.getAgent(self.sessID, agentID)
+                agent = daemon.getAgent(self, agentID)
                 if((bestAgent is None) or (bestAgent.fitness < agent.fitness)):
                     bestAgent = agent
                     
         except Exception as e:
-            print("gaDiscrete.getBestAgent failed with error as : {}".format(e))
+            print("gaDisc.getBestAgent failed with error as : {}".format(e))
             bestAgent = None
         
         return(bestAgent)
 
+    
     def getAverageFitness(self):
         """
         INPUT       : None
@@ -161,22 +149,23 @@ class Session:
         DESCRIPTION : Self explanatory from OUTPUT
 
         This function is used to get the average fitness value of
-        agents in the current session, this uses deamon functions
+        agents in the current session, this uses daemon functions
         getCurrGen, getAgent
         """
 
         try:
             totalFitness = 0
-            currGen = deamon.getCurrGen(self.sessID)
+            currGen = daemon.getCurrGen(self)
             for agentID in currGen:
-                totalFitness += deamon.getAgent(self.sessID, agentID).fitness
+                totalFitness += daemon.getAgent(self, agentID).fitness
 
-        except Exception:
-            print("gaDiscrete.getAverageFitness failed with error as : {}".format(e))
+        except Exception as e:
+            print("gaDisc.getAverageFitness failed with error as : {}".format(e))
             totalFitness = 0
 
         return(totalFitness/self.agentCount)
                 
+
     def createNextGen(self):
         """
         INPUT       : None
@@ -198,40 +187,46 @@ class Session:
         
         all the unselected agents are deleted (killed ?)
         now the current population is used to create new children
-        by calling a deamon function aSexualRep, which returns a
+        by calling a daemon function aSexualRep, which returns a
         new agentObj as a child when a parent agentObj is sent
         this new child-agentObj will inherit the fitness value and
-        slightly modified from its parent
+        slightly modified from its parent, (Needs optimisation).
         """
         
         try:
             if(self.survival > 0.95):
                 survivalPerc = 0.95
                 
-            numberOfSurvivors = int((random.uniform(self.survival, self.survival+0.05) * self.agentCount))                      
+            numberOfSurvivors = int(((self.survival+0.05*random.random()) * self.agentCount))                      
             if(numberOfSurvivors < 1):
                 numberOfSurvivors = 1
 
             listOfAgents = [(agent.agentID, agent.fitness) for agent in
-                            [deamon.getAgent(self.sessID, agentID) for agentID in
-                             deamon.getCurrGen(self.sessID)]]
+                            [daemon.getAgent(self, agentID) for agentID in
+                             daemon.getCurrGen(self)]]
 
             listOfAgents.sort(key=lambda x: x[1], reverse=True)
             unselected = listOfAgents[numberOfSurvivors:]
-            wildCardEntries = random.sample(unselected, int(len(unselected)*0.01))
+            # wildcard zero
+            wildCardEntries = random.sample(unselected, int(len(unselected)*0.00))
             finalUnselected = [agent for agent in unselected if(agent not in wildCardEntries)]
 
             for agent in finalUnselected:
-                deamon.deleteAgent(self.sessID, agent[0])
+                daemon.deleteAgent(self, agent[0])
 
-            currPop = list(deamon.getCurrGen(self.sessID))
+            currPop = list(daemon.getCurrGen(self))
             itr = 0
-            while(len(deamon.getCurrGen(self.sessID)) < self.agentCount):
-                self.updateAgent(deamon.aSexualRep(self, self.getAgent(currPop[itr])))
+            while(len(daemon.getCurrGen(self)) < self.agentCount):
+                if(self.generateMode == 'A'):
+                    self.updateAgent(daemon.aSexualRep(self, self.getAgent(currPop[itr])))
+                else:
+                    self.updateAgent(daemon.sexualRep(self, self.getAgent(currPop[itr]),
+                                                      self.getAgent(currPop[(itr + 1)%(len(currPop))])))
                 itr = (itr + 1)%(len(currPop))
-
+        
         except Exception as e:
-            print("gaDiscrete.createNextGen failed with error as : {}".format(e))
+            print("gaDisc.createNextGen failed with error as : {}".format(e))
+
             
     def resetEnv(self):
         """
@@ -242,24 +237,30 @@ class Session:
         
         This function effectively resets the session to the
         "Just created form" all the meta data files along with previous
-        agents are deleted (killed ?). Uses the deamon (pun intended)
+        agents are deleted (killed ?). Uses the daemon (pun intended)
         functions setSession which registeres this session for future use
         """
-        
-        try:            
-            sessDir = deamon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
-            shutil.rmtree(sessDir+"/smtf")
-            shutil.rmtree(sessDir+"/dnaPool")
-            os.mkdir(sessDir+"/smtf")
-            os.mkdir(sessDir+"/dnaPool")
-            fp = open(sessDir+"/smtf/LOCK.smtf", "w")
-            fp.write(str(0))
-            fp.close()
-            deamon.setSession(self)
-            
-        except Exception as e:
-            print("gaDiscrete.resetEnv failed with error as : {}".format(e))
-            
+
+        if(self.mode == 'SAFE'):
+            try:            
+                sessDir = daemon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
+                shutil.rmtree(sessDir+"/smtf")
+                shutil.rmtree(sessDir+"/dnaPool")
+                os.mkdir(sessDir+"/smtf")
+                os.mkdir(sessDir+"/dnaPool")
+                fp = open(sessDir+"/smtf/LOCK.smtf", "w")
+                fp.write(str(0))
+                fp.close()
+                daemon.setSession(self)
+                
+            except Exception as e:
+                print("gaDisc.resetEnv failed with error as : {}".format(e))
+        else:
+            self.currGen       = set()
+            self.agentBasket   = dict()
+            self.nextAgentID   = 0
+            self.lock          = 0
+
 
     def delete(self):
         """
@@ -270,13 +271,17 @@ class Session:
         
         As the name of the function says, this function deletes all of
         the meta data generated by this session and effectively deletes
-        this session uses the deamon data GA_UTIL_DIR
+        this session uses the daemon data GA_UTIL_DIR
         """
 
-        try:
-            sessDir = deamon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
-            shutil.rmtree(sessDir)
-            del self
+        if(self.mode == 'SAFE'):
+            try:
+                sessDir = daemon.GA_UTIL_DIR + "/utilFiles/tmp" + str(self.sessID)
+                shutil.rmtree(sessDir)
+                del self
             
-        except Exception as e:
-            print("gaDiscrete.delete failed with error as : {}".format(e))
+            except Exception as e:
+                print("gaDisc.delete failed with error as : {}".format(e))
+        else:
+            del self
+
